requirements.txt
asgiref==3.8.1
Django==4.2.13
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
drf-extra-fields==3.6.0
filetype==1.2.0
install==1.3.5
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
